Added Group Policy Update functionality and fixed test code
Added new classes and methods to perform Group Policy updates. Added a new method in `AdmxExtensions.cs` to get the registry keys for the local machine and the current user, and added a reference to the `Microsoft.Win32.Registry` package in `AdmxParser.csproj` to do this. New classes for performing Group Policy updates have been added to `UpdatePolicy.cs` and `NativeMethods.cs`. A new test class has also been added to `UpdatePolicyTest.cs` to test these changes.

In the existing code, the test class constructor in `AdmxDirectoryTest.cs` has been removed and the indentation of the namespace string in `Helpers.cs` has been corrected.